+++ go/clustermetadata/topo/cell_location.go
@@ -1,4 +1,4 @@
-// Copyright 2025 The Multigres Authors
+// Copyright 2025 The Multigres Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@ import (
 	"errors"
 	"path"
 
-	clustermetadatapb "github.com/multigres/multigres/pb/clustermetadata"
+	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
 
 	"google.golang.org/protobuf/proto"
 )
