+++ go/clustermetadata/topo/conn.go
@@ -1,4 +1,4 @@
-// Copyright 2025 The Multigres Authors
+// Copyright 2025 The Multigres Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
