+++ go/clustermetadata/topo/memorytopo/file.go
@@ -1,3 +1,17 @@
+// Copyright 2025 The Multigres Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package memorytopo
 
 import (
@@ -6,9 +20,9 @@ import (
 	"path"
 	"strings"
 
-	"github.com/multigres/multigres/pb/mtrpc"
 	"github.com/multigres/multigres/go/clustermetadata/topo"
 	"github.com/multigres/multigres/go/mterrors"
+	"github.com/multigres/multigres/go/pb/mtrpc"
 )
 
 // Below we implement the File methods that are part of the Conn interface:
