+++ go/clustermetadata/topo/memorytopo/store.go
@@ -1,3 +1,17 @@
+// Copyright 2025 The Multigres Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 // Package memorytopo contains an implementation of the topo.Factory /
 // topo.Conn interfaces based on an in-memory tree of data.
 // It is constructed with an immutable set of cells.
@@ -16,7 +30,7 @@ import (
 
 	"github.com/multigres/multigres/go/clustermetadata/topo"
 
-	clustermetadatapb "github.com/multigres/multigres/pb/clustermetadata"
+	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
 )
 
 const (
